src/main/java/com/example/tomasz1452/model/projection/GroupReadModel.java
package com.example.tomasz1452.model.projection;

import com.example.tomasz1452.model.Task;
import com.example.tomasz1452.model.TaskGroup;

import java.time.LocalDateTime;
import java.util.Objects;
import java.util.Set;
import java.util.stream.Collectors;

public class GroupReadModel {
    private int id;
    private String description;
    /**
     * Deadline from the latest task in group
     */
    private LocalDateTime deadline;
    private Set<GroupTaskReadModel> tasks;

    public GroupReadModel(TaskGroup source){
        id = source.getId();
        description = source.getDescription();
        source.getTasks().stream()
                .map(Task::getDeadline)
                .filter(Objects::nonNull) // To odfiltrowanie dodaliśmy w wyniku testu w GroupReadModelTest.
                .max(LocalDateTime::compareTo)
                .ifPresent((date -> deadline = date));
//        Tutaj pobieramy ostatnie zadanie z grupy o najdalszej dacie wykonania.

        tasks = source.getTasks().stream()
                .map(GroupTaskReadModel::new)
                .collect(Collectors.toSet());
//        Tutaj jakbyśmy zebrali cała grupę i zrzutowali ją do typu Set.
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public LocalDateTime getDeadline() {
        return deadline;
    }

    public void setDeadline(LocalDateTime deadline) {
        this.deadline = deadline;
    }

    public Set<GroupTaskReadModel> getTasks() {
        return tasks;
    }

    public void setTasks(Set<GroupTaskReadModel> tasks) {
        this.tasks = tasks;
    }

    public void setId(int id) {
        this.id = id;
    }

    public int getId() {
        return id;
    }
}
